requirements.txt
Flask
pyotp
pillow
cryptography
bcrypt
python-dotenv
qrcode
bleach
flask-limiter